pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx2txt>=0.9",
    "google-generativeai>=0.8.5",
    "openai>=1.76.0",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "streamlit>=1.44.1",
]
